requirements.txt
bidict==0.22.0
blinker==1.5
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.2
greenlet==2.0.0.post0
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql==0.0.3
mysql-connector-python==8.0.31
mysqlclient==2.1.1
numpy==1.23.4
packaging==21.3
pluggy==1.0.0
protobuf==3.20.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.6
requests==2.28.1
six==1.11.0
SQLAlchemy==1.4.42
tomli==2.0.1
urllib3==1.26.12
Werkzeug==2.2.0
WTForms==3.0.1
zipp==3.8.0
